feat: Add support for (custom) types (#498)
* Add support for (custom) types

This adds the `Types` field to the ResourceGraphDefinition CRD. It
contains a map of names to type definitions, where the definitions
follow the simple schema.

The defined types are then loaded into the transformer as predefined
types. Appearances in the spec are then properly resolved.

An example for usage would be:

``` yaml
apiVersion: kro.run/v1alpha1
kind: ResourceGraphDefinition
metadata:
  name: service-entry-dns
spec:
  schema:
    apiVersion: v1alpha1
    kind: ServiceEntryDNS
    spec:
      ports: "[]port | required=true"
    types:
      port:
         name: string
         protocol: string
         number: integer | required
```

Known issues: Currently there is no support for references of custom
types within the definition of a custom type!

Fixes #144

* Add documentation for (custom) types
* Fix (custom) simple types with required marker

If a simple type was used with the required marker:

``` yaml
types:
  myType: string | required=true description="my description"
spec:
  myValue: myType
```
the required marker was lost, while the description was set properly.
This happened as in the applyMarkers method of the transformer the
required information is set to the parent.

While loading the predefined types this parent is just a shell object,
that is later read to fill the list of predefined types and then
discarded.

Now the information on required types is captured in the
`predefinedType` type and then later used to set the parent object
required fields when handling predefined types.

* Fix typo in documentation
* Remove unused code
* Fix and improve test for predefined types
* Simplify code